separated login and register
