--- pig/Views/AuthView.php
@@ -1,6 +1,0 @@
-<?php
-class AuthView {
-    public function render() {
-        include __DIR__ . '/AuthView.tpl';
-    }
-}
