+++ pig/Views/LoginView.php
@@ -1,0 +1,6 @@
+<?php
+class LoginView {
+    public function render() {
+        include __DIR__ . '/Loginview.tpl';
+    }
+}
