+++ pig/Views/RegisterView.php
@@ -1,0 +1,6 @@
+<?php
+class LoginView {
+    public function render() {
+        include __DIR__ . '/Registerview.tpl';
+    }
+}
