+++ pig/index.php
@@ -1,10 +1,25 @@
 <?php
-require_once __DIR__ . '/Controllers/AuthController.php';
+require_once __DIR__ . '/../vendor/autoload.php';
 
-$conn = oci_connect('student', 'STUDENT', 'localhost/XE');
+$env = Dotenv\Dotenv::createImmutable(__DIR__ ."/../");
+$env->load();
+
+require_once __DIR__ . '/Controllers/LoginController.php';
+require_once __DIR__ . '/Controllers/RegisterController.php';
+
+
+
+$conn = oci_connect($_ENV['DB_NAME'], $_ENV['DB_PASSWORD'], $_ENV['DB_CONNECTION_STRING']);
 if (!$conn) {
     die("Eroare la conectarea la baza de date!");
 }
 
-$authController = new AuthController($conn);
-$authController->design();
+$page = $_GET['page'] ?? 'login';
+
+if ($page === 'register') {
+    $controller = new RegisterController($conn);
+    $controller->design();
+} else {
+    $controller = new LoginController($conn);
+    $controller->design();
+}
