+++ pig/js/register.js
@@ -6,9 +6,9 @@ window.showPassword = function(id) {
 };
 
 document.addEventListener('DOMContentLoaded', () => {
-
     const registerForm = document.getElementById('registerForm');
     const registerMessage = document.getElementById('registerMessage');
+
     registerForm.addEventListener('submit', async (e) => {
         e.preventDefault();
         registerMessage.textContent = 'Se încarcă...';
@@ -20,36 +20,17 @@ document.addEventListener('DOMContentLoaded', () => {
             method: 'POST',
             body: formData
         });
+
         const data = await response.json();
         registerMessage.textContent = data.message;
         if (data.success) registerForm.reset();
     });
 
-    const loginForm = document.getElementById('loginForm');
-    const loginMessage = document.getElementById('loginMessage');
-    loginForm.addEventListener('submit', async (e) => {
-        e.preventDefault();
-        loginMessage.textContent = 'Se încarcă...';
-
-        const formData = new FormData(loginForm);
-        formData.append('action', 'login');
-
-        const response = await fetch('/ProiectWEB_Darie_Mihnea_Stefan_2A2_Ciurariu_Raluca_Iuliana_2A2/pig/Controllers/handler.php', {
-            method: 'POST',
-            body: formData
-        });
-        const data = await response.json();
-        loginMessage.textContent = data.message;
-        if (data.success) loginForm.reset();
-    });
-
-    const showPasswordButton=document.querySelectorAll('button[data-show-password]');
-    showPasswordButton.forEach(button=>{
+    const showPasswordButtons = document.querySelectorAll('button[data-show-password]');
+    showPasswordButtons.forEach(button => {
         button.addEventListener('click', () => {
             const inputId = button.dataset.showPassword;
             showPassword(inputId);
         });
-    }
-    )
-
+    });
 });
